04_01_24 Adding helper files for plotting and made changes to priority and serial backend

+++ libraries/AP_Proximity/AP_Proximity_Backend_Serial.h
@@ -18,7 +18,7 @@ public:
     static bool detect(uint8_t serial_instance);
 
 protected:
-    virtual uint16_t rxspace() const { return 0; };
+    virtual uint32_t rxspace() const { return 0; };
 
     AP_HAL::UARTDriver *_uart;              // uart for communicating with sensor
 };
